Moved flight path component to map overlays file
